requirements.txt
crewai
crewai-tools
composio-crewai
streamlit
langchain
langchain-openai
google-search-results 
python-dotenv
firecrawl-py
composio_core
composio_openai 
PyGithub
requests
unstructured
pypdf
markdown
xmltodict
PyGithub
unstructured
pypdf
markdown
xmltodict
chromadb
embedchain
beautifulsoup4
readability-lxml
youtube-search-python
google-api-python-client

